% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_caret.R
\name{print.wrap_caret}
\alias{print.wrap_caret}
\title{Print method for wrap_caret objects}
\usage{
\method{print}{wrap_caret}(x, ...)
}
\arguments{
\item{x}{Object from wrap_caret}

\item{...}{Additional arguments}
}
\description{
Print method for wrap_caret objects
}
