% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_caret.R
\name{predict.wrap_caret}
\alias{predict.wrap_caret}
\title{Predict method for mlS3 caret wrapper}
\usage{
\method{predict}{wrap_caret}(object, newx, type = NULL, ...)
}
\arguments{
\item{object}{Object from wrap_caret}

\item{newx}{New features (matrix or data frame)}

\item{type}{Prediction type: "raw" (default), "class", "prob", or NULL}

\item{...}{Additional arguments to caret::predict.train}
}
\value{
Vector or matrix of predictions
}
\description{
Predict method for mlS3 caret wrapper
}
